refactor(test:streaming-parity): reduce cognitive complexity of assert_fixture_parity
Extract comparison and error checking logic into helper functions in streaming_test_support to reduce cognitive complexity from 23 to under 15. Split monolithic assert_fixture_parity into focused helpers: check_conversion_errors and check_output_comparison. Also consolidate duplicate known_differences module imports.
