+++ components/rust-converter/tests/streaming_parity.rs
@@ -13,8 +13,6 @@
 
 #![cfg(feature = "streaming")]
 
-#[path = "known_differences.rs"]
-mod known_differences;
 #[path = "streaming_test_support.rs"]
 mod streaming_test_support;
 
@@ -23,28 +21,16 @@ use std::path::Path;
 use std::process::Command;
 use std::time::Instant;
 
-use known_differences::{KnownDifferences, OutputDifference};
+use streaming_test_support::known_differences::KnownDifferences;
 use nginx_markdown_converter::error::ConversionError;
+
 use streaming_test_support::{
+    check_conversion_errors, check_output_comparison,
     convert_full_buffer, convert_streaming_chunked, convert_streaming_single,
     default_streaming_budget, default_streaming_options, discover_html_fixtures,
     evidence_output_path, fixture_relative_name, known_differences_path,
     normalize_whitespace_tokens, read_fixture, read_fixture_meta,
 };
-
-#[derive(Debug)]
-enum ComparisonResult {
-    Identical,
-    KnownDifference {
-        diff_id: String,
-        description: String,
-    },
-    Divergence {
-        full_buffer_output: String,
-        streaming_output: String,
-        diff: String,
-    },
-}
 
 fn discover_fixtures(corpus_dir: &Path) -> Vec<std::path::PathBuf> {
     discover_html_fixtures(corpus_dir)
@@ -89,97 +75,6 @@ fn convert_streaming_chunked_entry(
     Ok(run.markdown)
 }
 
-fn compare_outputs(
-    fixture_name: &str,
-    full_buffer: &str,
-    streaming: &str,
-    known_diffs: &KnownDifferences,
-) -> ComparisonResult {
-    if full_buffer == streaming {
-        return ComparisonResult::Identical;
-    }
-
-    let mut diff = unified_diff_summary(full_buffer, streaming);
-    if normalize_whitespace_tokens(full_buffer) == normalize_whitespace_tokens(streaming) {
-        diff = format!("whitespace-only-parity-drift\n{diff}");
-    }
-
-    let output = OutputDifference {
-        full_buffer,
-        streaming,
-        diff: &diff,
-    };
-
-    if let Some(entry) = known_diffs.matches(fixture_name, &output) {
-        return ComparisonResult::KnownDifference {
-            diff_id: entry.id.clone(),
-            description: entry.description.clone(),
-        };
-    }
-
-    ComparisonResult::Divergence {
-        full_buffer_output: full_buffer.to_string(),
-        streaming_output: streaming.to_string(),
-        diff,
-    }
-}
-
-fn unified_diff_summary(expected: &str, actual: &str) -> String {
-    const MAX_DIFF_LINES: usize = 3;
-
-    if expected == actual {
-        return "<identical>".to_string();
-    }
-
-    let lhs: Vec<&str> = expected.lines().collect();
-    let rhs: Vec<&str> = actual.lines().collect();
-    let shared = lhs.len().min(rhs.len());
-    let mut diffs = Vec::new();
-
-    for idx in 0..shared {
-        if lhs[idx] != rhs[idx] {
-            diffs.push(format!(
-                "line {} differs\n- {}\n+ {}",
-                idx + 1,
-                lhs[idx],
-                rhs[idx]
-            ));
-            if diffs.len() == MAX_DIFF_LINES {
-                break;
-            }
-        }
-    }
-
-    if !diffs.is_empty() {
-        return diffs.join("\n");
-    }
-
-    if lhs.len() != rhs.len() {
-        return format!(
-            "line count differs (full-buffer={}, streaming={})",
-            lhs.len(),
-            rhs.len()
-        );
-    }
-
-    "outputs differ but first mismatch not localized".to_string()
-}
-
-fn is_known_runtime_difference(
-    fixture_name: &str,
-    marker: &str,
-    detail: &str,
-    known_diffs: &KnownDifferences,
-) -> bool {
-    let diff = format!("{marker}\n{detail}");
-    let output = OutputDifference {
-        full_buffer: "",
-        streaming: "",
-        diff: &diff,
-    };
-    known_diffs.matches(fixture_name, &output).is_some()
-}
-
 fn assert_fixture_parity(path: &Path, known_diffs: &KnownDifferences) -> Result<(), String> {
     let fixture_name = fixture_relative_name(path);
     let meta = read_fixture_meta(path);
@@ -213,100 +108,18 @@ fn assert_fixture_parity(path: &Path, known_diffs: &KnownDifferences) -> Result<
     let single = convert_streaming_single_entry(&html, Some(&content_type), budget.clone());
     let chunked = convert_streaming_chunked_entry(&html, &chunk_sizes, Some(&content_type), budget);
 
-    match (&single, &chunked) {
-        (
-            Err(ConversionError::StreamingFallback { .. }),
-            Err(ConversionError::StreamingFallback { .. }),
-        ) => {
-            if !meta.expected_fallback {
-                if is_known_runtime_difference(
-                    &fixture_name,
-                    "streaming-fallback",
-                    "both-single-and-chunked",
-                    known_diffs,
-                ) {
-                    return Ok(());
-                }
-                return Err(format!(
-                    "{fixture_name}: unexpected fallback (meta expected_fallback=false)"
-                ));
-            }
-            return Ok(());
-        }
-        (Err(err), _) if !matches!(err, ConversionError::StreamingFallback { .. }) => {
-            if is_known_runtime_difference(
-                &fixture_name,
-                "streaming-error",
-                &err.to_string(),
-                known_diffs,
-            ) {
-                return Ok(());
-            }
-            return Err(format!("{fixture_name}: streaming single failed: {err}"));
-        }
-        (_, Err(err)) if !matches!(err, ConversionError::StreamingFallback { .. }) => {
-            if is_known_runtime_difference(
-                &fixture_name,
-                "streaming-error",
-                &err.to_string(),
-                known_diffs,
-            ) {
-                return Ok(());
-            }
-            return Err(format!("{fixture_name}: streaming chunked failed: {err}"));
-        }
-        (Err(ConversionError::StreamingFallback { .. }), Ok(_))
-        | (Ok(_), Err(ConversionError::StreamingFallback { .. })) => {
-            if is_known_runtime_difference(
-                &fixture_name,
-                "streaming-fallback-mismatch",
-                "single-vs-chunked",
-                known_diffs,
-            ) {
-                return Ok(());
-            }
-            return Err(format!(
-                "{fixture_name}: single/chunked fallback behavior mismatch"
-            ));
-        }
-        _ => {}
+    let should_compare = check_conversion_errors(
+        &fixture_name, &meta, &single, &chunked, known_diffs,
+    )?;
+
+    if !should_compare {
+        return Ok(());
     }
 
-    if meta.expected_fallback {
-        return Err(format!(
-            "{fixture_name}: expected fallback but streaming produced Markdown"
-        ));
-    }
+    let single = single.unwrap();
+    let chunked = chunked.unwrap();
 
-    let single = single.unwrap_or_else(|_| unreachable!());
-    let chunked = chunked.unwrap_or_else(|_| unreachable!());
-
-    for (label, streaming_output) in [("single", single), ("chunked", chunked)] {
-        match compare_outputs(&fixture_name, &full_buffer, &streaming_output, known_diffs) {
-            ComparisonResult::Identical => {}
-            ComparisonResult::KnownDifference {
-                diff_id,
-                description,
-            } => {
-                if !meta.known_diff_ids.iter().any(|id| id == &diff_id) {
-                    return Err(format!(
-                        "{fixture_name}: matched known diff {diff_id} ({description}) but fixture metadata does not list it"
-                    ));
-                }
-            }
-            ComparisonResult::Divergence {
-                full_buffer_output,
-                streaming_output,
-                diff,
-            } => {
-                return Err(format!(
-                    "{fixture_name}: {label} output divergence\n{diff}\n--- full-buffer ---\n{full_buffer_output}\n--- streaming ---\n{streaming_output}"
-                ));
-            }
-        }
-    }
-
-    Ok(())
+    check_output_comparison(&fixture_name, &full_buffer, &single, &chunked, &meta, known_diffs)
 }
 
 fn generate_large_fixtures_if_needed() {
